Move function to the ncc taskfile
